+++ decomp/ST.exe/functions/0070B770/decomp.c
@@ -25,7 +25,7 @@ int __cdecl FUN_0070b770(AnonShape_0070B770_3BA0472D *param_1,int param_2)
           g_currentExceptionFrame = pIVar4;
           return local_8;
         }
-        iVar2 = FUN_006d1a10((undefined4 *)0x0,0,0,0,(int)param_1,0,0,0,param_1->field_0004,
+        iVar2 = FUN_006d1a10((undefined4 *)0x0,0,0,0,param_1,0,0,0,param_1->field_0004,
                              param_1->field_0008);
         g_currentExceptionFrame = pIVar4;
         return iVar2;
@@ -40,7 +40,7 @@ int __cdecl FUN_0070b770(AnonShape_0070B770_3BA0472D *param_1,int param_2)
         g_currentExceptionFrame = local_90.previous;
         return local_8;
       }
-      iVar2 = FUN_006d1a50((undefined4 *)0x0,0,0,0,(int)param_1,0,0,0,param_1->field_0004,
+      iVar2 = FUN_006d1a50((undefined4 *)0x0,0,0,0,param_1,0,0,0,param_1->field_0004,
                            param_1->field_0008);
       g_currentExceptionFrame = local_90.previous;
       return iVar2;
@@ -63,7 +63,7 @@ int __cdecl FUN_0070b770(AnonShape_0070B770_3BA0472D *param_1,int param_2)
           g_currentExceptionFrame = local_4c.previous;
           return local_8;
         }
-        iVar2 = FUN_006d1a90((undefined4 *)0x0,0,0,0,(int)param_1,0,0,0,param_1->field_0004,
+        iVar2 = FUN_006d1a90((undefined4 *)0x0,0,0,0,param_1,0,0,0,param_1->field_0004,
                              param_1->field_0008);
         g_currentExceptionFrame = local_4c.previous;
         return iVar2;
